backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
starlette==0.36.3        # Required by FastAPI and session middleware

# Configuration & Environment
python-dotenv==1.0.1
pydantic[email]==2.11.0
pydantic-settings==2.2.1
email-validator==2.1.0   # Required by pydantic EmailStr

# Database
sqlalchemy==2.0.36       # Updated for Python 3.13 compatibility
psycopg==3.2.3          # PostgreSQL adapter (pure Python, works better on Windows)
alembic==1.13.2          # Database migrations

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4   # Password hashing
Authlib==1.3.1           # OAuth
httpx==0.27.0            # Required by authlib for OAuth

# Session middleware
itsdangerous==2.2.0      # For session signing